+++ services/preview/appplatform/mgmt/2020-11-01-preview/appplatform/models.go
@@ -238,14 +238,14 @@ type AppResourceProperties struct {
 	Fqdn *string `json:"fqdn,omitempty"`
 	// HTTPSOnly - Indicate if only https is allowed.
 	HTTPSOnly *bool `json:"httpsOnly,omitempty"`
-	// EnableEndToEndTLS - Indicate if end to end TLS is enabled.
-	EnableEndToEndTLS *bool `json:"enableEndToEndTLS,omitempty"`
 	// CreatedTime - READ-ONLY; Date time when the resource is created
 	CreatedTime *date.Time `json:"createdTime,omitempty"`
 	// TemporaryDisk - Temporary disk settings
 	TemporaryDisk *TemporaryDisk `json:"temporaryDisk,omitempty"`
 	// PersistentDisk - Persistent disk settings
 	PersistentDisk *PersistentDisk `json:"persistentDisk,omitempty"`
+	// EnableEndToEndTLS - Indicate if end to end TLS is enabled.
+	EnableEndToEndTLS *bool `json:"enableEndToEndTLS,omitempty"`
 }
 
 // MarshalJSON is the custom marshaler for AppResourceProperties.
@@ -263,14 +263,14 @@ func (arp AppResourceProperties) MarshalJSON() ([]byte, error) {
 	if arp.HTTPSOnly != nil {
 		objectMap["httpsOnly"] = arp.HTTPSOnly
 	}
-	if arp.EnableEndToEndTLS != nil {
-		objectMap["enableEndToEndTLS"] = arp.EnableEndToEndTLS
-	}
 	if arp.TemporaryDisk != nil {
 		objectMap["temporaryDisk"] = arp.TemporaryDisk
 	}
 	if arp.PersistentDisk != nil {
 		objectMap["persistentDisk"] = arp.PersistentDisk
+	}
+	if arp.EnableEndToEndTLS != nil {
+		objectMap["enableEndToEndTLS"] = arp.EnableEndToEndTLS
 	}
 	return json.Marshal(objectMap)
 }
@@ -2397,6 +2397,8 @@ type MetricDimension struct {
 	Name *string `json:"name,omitempty"`
 	// DisplayName - Localized friendly display name of the dimension
 	DisplayName *string `json:"displayName,omitempty"`
+	// ToBeExportedForShoebox - Whether this dimension should be included for the Shoebox export scenario
+	ToBeExportedForShoebox *bool `json:"toBeExportedForShoebox,omitempty"`
 }
 
 // MetricSpecification specifications of the Metrics for Azure Monitoring
@@ -2421,6 +2423,8 @@ type MetricSpecification struct {
 	FillGapWithZero *bool `json:"fillGapWithZero,omitempty"`
 	// Dimensions - Dimensions of the metric
 	Dimensions *[]MetricDimension `json:"dimensions,omitempty"`
+	// SourceMdmNamespace - Name of the MDM namespace. Optional.
+	SourceMdmNamespace *string `json:"sourceMdmNamespace,omitempty"`
 }
 
 // MonitoringSettingProperties monitoring Setting properties payload
@@ -2646,10 +2650,33 @@ type OperationDetail struct {
 	IsDataAction *bool `json:"isDataAction,omitempty"`
 	// Display - Display of the operation
 	Display *OperationDisplay `json:"display,omitempty"`
+	// ActionType - READ-ONLY; Enum. Indicates the action type. "Internal" refers to actions that are for internal only APIs. Possible values include: 'Internal'
+	ActionType ActionType `json:"actionType,omitempty"`
 	// Origin - Origin of the operation
 	Origin *string `json:"origin,omitempty"`
 	// Properties - Properties of the operation
 	Properties *OperationProperties `json:"properties,omitempty"`
+}
+
+// MarshalJSON is the custom marshaler for OperationDetail.
+func (od OperationDetail) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	if od.Name != nil {
+		objectMap["name"] = od.Name
+	}
+	if od.IsDataAction != nil {
+		objectMap["isDataAction"] = od.IsDataAction
+	}
+	if od.Display != nil {
+		objectMap["display"] = od.Display
+	}
+	if od.Origin != nil {
+		objectMap["origin"] = od.Origin
+	}
+	if od.Properties != nil {
+		objectMap["properties"] = od.Properties
+	}
+	return json.Marshal(objectMap)
 }
 
 // OperationDisplay operation display payload