+++ services/preview/appplatform/mgmt/2020-11-01-preview/appplatform/CHANGELOG.md
@@ -1,2 +1,20 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Constants
+
+1. ActionType.Internal
+
+### New Funcs
+
+1. OperationDetail.MarshalJSON() ([]byte, error)
+1. PossibleActionTypeValues() []ActionType
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. MetricDimension.ToBeExportedForShoebox
+1. MetricSpecification.SourceMdmNamespace
+1. OperationDetail.ActionType