+++ services/preview/appplatform/mgmt/2021-06-01-preview/appplatform/CHANGELOG.md
@@ -1,2 +1,19 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Constants
+
+1. ActionType.ActionTypeInternal
+
+### New Funcs
+
+1. OperationDetail.MarshalJSON() ([]byte, error)
+1. PossibleActionTypeValues() []ActionType
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. MetricSpecification.SourceMdmNamespace
+1. OperationDetail.ActionType