+++ services/preview/appplatform/mgmt/2021-09-01-preview/appplatform/CHANGELOG.md
@@ -1,2 +1,20 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Constants
+
+1. ActionType.ActionTypeInternal
+1. ProvisioningState.ProvisioningStateStarting
+1. ProvisioningState.ProvisioningStateStopping
+
+### New Funcs
+
+1. OperationDetail.MarshalJSON() ([]byte, error)
+1. PossibleActionTypeValues() []ActionType
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. OperationDetail.ActionType